requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
marshmallow==3.20.1
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
requests==2.31.0
email-validator==2.1.0.post1
flask-limiter==3.5.0
filelock==3.13.1